pyproject.toml
[tool.poetry]
name = "aidevs3"
version = "0.1.0"
description = "Files for AIDevs_3 course"
authors = ["Dominik Jeziorski <d.jeziorski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
loguru = "^0.7.2"
groq = "^0.11.0"
ollama = "^0.3.3"
llama-index = "^0.11.21"
langfuse = "^2.53.9"
sympy = "^1.13.1"
pydub = "^0.25.1"
whisper = "*"
paddleocr = "^2.9.1"
markdown2 = "^2.5.1"
pathvalidate = "^3.2.1"
mpu = "^0.23.1"
tabulate = "^0.9.0"
qdrant-client = "^1.12.1"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
